% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actor_concept_el}
\alias{actor_concept_el}
\title{Actor-concept edgelist}
\format{
A data frame with 140 rows and 2 variables
\describe{
  \item{sender}{Actors}
  \item{receiver}{Activities}
  ...
}
}
\source{
Reussebene survey
}
\usage{
actor_concept_el
}
\description{
A dataset, specifying actor-activity ties. Thus,
every row specifies that an actor (sender) partakes in a specific activity (receiver)
}
\keyword{datasets}
