% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actor_el}
\alias{actor_el}
\title{Actor edgelist}
\format{A data frame with n rows and n variables:
\describe{
  \item{sender}{Actors who indicated collaboration}
  \item{receiver}{Target actor of collaboration indicated by sender}
  ...
}}
\source{
survey
}
\usage{
actor_el
}
\description{
A dataset specifying collaboration links between actors. Every record indicates a tie between actors.
}
\keyword{datasets}
