% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actors}
\alias{actors}
\title{Actor dataframe}
\format{A data frame with n rows and n variables:
\describe{
  \item{actor}{tbd}
  ...
}}
\source{
survey
}
\usage{
actors
}
\description{
A dataset
}
\keyword{datasets}
