% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cld_concepts}
\alias{cld_concepts}
\title{CLD concepts dataframe}
\format{A data frame with n rows and n variables:
\describe{
  \item{concept}{tbd}
  \item{type}{tbd}
  \item{english}{tbd}
  ...
}}
\source{
survey
}
\usage{
cld_concepts
}
\description{
A dataset
}
\keyword{datasets}
