% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cld_el}
\alias{cld_el}
\title{CLD edgelist dataframe}
\format{A data frame with n rows and n variables:
\describe{
  \item{sender}{Sender concept}
  \item{receiver}{Target concept}
  \item{sign}{Type of relation (negative/ positive)}
  ...
}}
\source{
survey
}
\usage{
cld_el
}
\description{
A dataset of links between causal loop diagraming concepts. Every record indicates a link
}
\keyword{datasets}
