% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{count_motifs}
\alias{count_motifs}
\title{Count multi-level motifs}
\usage{
count_motifs(net, type_attr, number_nodes, motif)
}
\arguments{
\item{net}{A statnet network object with a node attribute specifying the level of each node}

\item{type_attr}{character vector specifying the attribute name where level information is stored in statnet object}

\item{number_nodes}{The number of nodes in the motif to be counted (3 or 4 are possible)}

\item{motif}{The motif name. Use show_3_motifs() or show_4_motifs() to see possible motifs}
}
\value{
Integer giving the motif count
}
\description{
Count multi-level motifs
}
