% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{critical_dyads}
\alias{critical_dyads}
\title{List critical dyads}
\usage{
critical_dyads(net, motif, type_attr = c("sesType"), level = -1)
}
\arguments{
\item{motif}{motif identifier}

\item{type_attr}{character vector specifying the attribute name where level
information is stored in statnet object. The attribute should be a binary
vector. 1 indicates a "social" node and 0 indicates a "non-social" node.}

\item{level}{level of the dyads which shall be considered, or -1 if the level
shall be determined automatically.}

\item{g}{statnet network object}
}
\value{
data frame with three columns, listing edges and their contribution
  to motifs described by the motif identifier in descending order
}
\description{
Critical dyads are edges on a specified level which break motifs by being
removed from the network.
}
\examples{
identify_gaps(ml_net, list('1,2[I.C]'))
}
