% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{directed_dummy_net}
\alias{directed_dummy_net}
\title{Two-level directed network dummy example}
\format{
igraph network object
}
\source{
Dummy data
 \url{https://gitlab.com/t.seppelt/sesmotifanalyser/-/tree/master/test/data},
 \url{https://github.com/marioangst/motifr/blob/master/notes/directed_dummy_net.R}
}
\usage{
directed_dummy_net
}
\description{
Simple igraph network object based on dummy data
}
\examples{
plot_mnet(directed_dummy_net)
}
\keyword{datasets}
