% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_net}
\alias{dummy_net}
\title{Three-level network dummy example}
\format{
Statnet network object with 60 nodes and 1035 edges on three levels.
The network contains two variables to describe nodes/ vertices.
\describe{
  \item{vertex.names}{node labes}
  \item{sesType}{Categorical variable specifying network levels for every node (levels are 0,1 and 2)}
  ...
}
}
\source{
Dummy data
}
\usage{
dummy_net
}
\description{
A simple statnet network object based on dummy data.
}
\keyword{datasets}
