% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{exemplify_motif}
\alias{exemplify_motif}
\title{Returns an example for a motif specified by a motif identifier string}
\usage{
exemplify_motif(net, motif, type_attr = c("sesType"))
}
\arguments{
\item{motif}{motif identifier string for the motif}

\item{type_attr}{character vector specifying the attribute name where level
information is stored in statnet object. The attribute should be a binary
vector. 1 indicates a "social" node and 0 indicates a "non-social" node.}

\item{g}{statnet network object}
}
\value{
vector of nodes in the motif
}
\description{
Returns an example for a motif specified by a motif identifier string
}
\examples{
exemplify_motif(ml_net, motif = '1,2[I.C]')
}
\seealso{
show_motif
}
