% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_baseline_ml_graph.R
\name{generate_random_ml_adjmat}
\alias{generate_random_ml_adjmat}
\title{Generate a random multi-level adjacency matrix}
\usage{
generate_random_ml_adjmat(n_social_nodes, n_non_social_nodes, density = NULL)
}
\arguments{
\item{n_social_nodes}{the number of social nodes in the network}

\item{n_non_social_nodes}{the number of "non-social" nodes in the network}

\item{density}{density of total network, defaults to random}
}
\value{
labeled adjacency matrix
}
\description{
Generate a random multi-level adjacency matrix
}
