% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_baseline_ml_graph.R
\name{generate_random_ml_net}
\alias{generate_random_ml_net}
\title{Generate a random multi-level network (statnet network object)}
\usage{
generate_random_ml_net(
  n_social_nodes,
  n_non_social_nodes,
  density = NULL,
  to_py = FALSE
)
}
\arguments{
\item{n_social_nodes}{the number of social nodes in the network}

\item{n_non_social_nodes}{the number of "non-social" nodes in the network}

\item{density}{density of total network, defaults to random}

\item{to_py}{should the object be returned as a sma compatible python object}
}
\value{
either a statnet network object with an attribute "sesType" specifying node type or an equivalent python compatible object
}
\description{
Generate a random multi-level network (statnet network object)
}
