% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cld.R
\name{get_entry_from_mapping}
\alias{get_entry_from_mapping}
\title{Get a entry for a specifc mapping. Used in visualize_cld_graph mainly.}
\usage{
get_entry_from_mapping(entry, mapping, default)
}
\arguments{
\item{entry}{Entry to get mapping from}

\item{mapping}{Named list, providing mappings}

\item{default}{Default value to choose if entry is not in mapping.}
}
\value{

}
\description{
Get a entry for a specifc mapping. Used in visualize_cld_graph mainly.
}
