% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cld.R
\name{get_shortest_inv_distance}
\alias{get_shortest_inv_distance}
\title{Get the shortest inverse distance between two issues}
\usage{
get_shortest_inv_distance(input_graph, max_path, agg_target_vec)
}
\arguments{
\item{input_graph}{Input graph}

\item{max_path}{Integer setting the maximum distance to still consider}

\item{agg_target_vec}{Character vector specifying nodes between which inverse shortest distances are to be calculated}
}
\description{
Get the shortest inverse distance between two issues
}
