% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is.directed}
\alias{is.directed}
\title{Checks whether the given network is directed}
\usage{
is.directed(net)
}
\arguments{
\item{net}{the network}
}
\value{
whether the given network is directed
}
\description{
Placeholder function for the corresponding functions of the various supported
network formats. For example, this function calls
\code{network::is.directed()} on \code{network} objects and
\code{igraph::is.directed()} on \code{igraph} objects.
}
\examples{
is.directed(ml_net)
}
