% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{large_directed_dummy_net}
\alias{large_directed_dummy_net}
\title{Large two-level directed network dummy example}
\format{
network network object
}
\source{
Dummy data
 \url{https://gitlab.com/t.seppelt/sesmotifanalyser/-/tree/master/test/data}
}
\usage{
large_directed_dummy_net
}
\description{
Large two-level directed network dummy example
}
\examples{
plot_mnet(large_directed_dummy_net)
}
\keyword{datasets}
