% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{motifs_distribution}
\alias{motifs_distribution}
\title{Compute statistical properties of the distribution of motifs in a random
baseline}
\usage{
motifs_distribution(
  net,
  motifs,
  type_attr = c("sesType"),
  model = c("erdos_renyi"),
  level = -1,
  omit_total_result = TRUE
)
}
\arguments{
\item{motifs}{list of motif identifiers describing the motifs whose
distribution shall be analysed}

\item{model}{model to be used, either erdos_renyi or actors_choice, see sma
documentation}

\item{level}{additional parameter for actors_choice}

\item{omit_total_result}{whether total results shall be omitted}

\item{g}{statnet network object}

\item{typeAttr}{character vector specifying the attribute name where level
information is stored in statnet object. The attribute should be a binary
vector. 1 indicates a "social" node and 0 indicates a "non-social" node.}
}
\value{
data frame indexed by motif identifers and with rows expectation and
  variances
}
\description{
Compute statistical properties of the distribution of motifs in a random
baseline
}
