% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps_critical_edges.R
\name{plot_critical_dyads}
\alias{plot_critical_dyads}
\title{Plot critical dyads in network vizualization}
\usage{
plot_critical_dyads(
  net,
  motif,
  lvl_attr = c("sesType"),
  level = -1,
  cutoff = 2,
  subset_graph = "none",
  ...
)
}
\arguments{
\item{net}{Statnet network object}

\item{motif}{Motif to explore gaps in for}

\item{lvl_attr}{Node attribute specifiying level information}

\item{level}{Focal level for gap analysis}

\item{cutoff}{Cutoff point in contributions of an edge to the number of
motifs above which to analyse gaps}

\item{subset_graph}{Whether to subset the graph to only show nodes involved
in gaps. One of "none" (no subset, default), "partial" (only focal level is
subset) or "focal" (only focal level shown)}

\item{...}{list of additional parameters to be passed to plotting function
(see ?motifr::plot_mnet), eg. label = TRUE}
}
\value{
A plot of gaps, sized by weight in a multilevel network
}
\description{
Note that this only works for undirected graphs. Regardless of whether the
input graph is directed it is treated as undirected graph.
}
\examples{
plot_critical_dyads(ml_net, "1,2[I.C]", level = -1)
plot_critical_dyads(ml_net, "1,2[I.C]",
  level = -1,
  subset_graph = "focal", cutoff = 4, label = TRUE
)
plot_critical_dyads(ml_net, "1,2[I.C]",
  level = -1,
  subset_graph = "partial", cutoff = 4, label = TRUE
)
}
