% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{plot_gaps}
\alias{plot_gaps}
\title{Plot gaps in network vizualization}
\usage{
plot_gaps(net, motif, type_attr = c("sesType"), level = -1, cutoff = 2, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot gaps in network vizualization
}
