% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mnet.R
\name{plot_mnet}
\alias{plot_mnet}
\title{Visualize a two-level network using ggraph}
\usage{
plot_mnet(net, type_attr = c("sesType"), layouts = list("kk", "kk"))
}
\arguments{
\item{layouts}{}
}
\value{

}
\description{
Visualize a two-level network using ggraph
}
