% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_based_on_codebook.R
\name{rename_based_on_codebook}
\alias{rename_based_on_codebook}
\title{Rename a vector based on a codebook}
\usage{
rename_based_on_codebook(input, codebook, rawvar, codevar)
}
\arguments{
\item{input}{Specify the vector to be recoded}

\item{codebook}{The data frame uses as a codebook for recoding}

\item{rawvar}{Character string giving the name of the column in the codebook containing references to the raw data}

\item{codevar}{Character string giving the name of the column in the codebook containing references to the coded data}
}
\value{
A named character vector of coded entries
}
\description{
Rename a vector based on a codebook
}
\examples{
animal_codebook <- data.frame(
animal = c("elephant", "pig", "lizard"),
family = c("mammal","mammal","reptile"))
vec <- sample(c("elephant", "pig", "lizard"), 50, TRUE)
rename_based_on_codebook(input = vec,
codebook = animal_codebook,
rawvar = "animal",
codevar = "family")
}
