% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cld.R
\name{save_cld_graph_as_svg}
\alias{save_cld_graph_as_svg}
\title{Save dot code generated in integrater as .svg file}
\usage{
save_cld_graph_as_svg(dot_code, file)
}
\arguments{
\item{dot_code}{Provide dot code generated through eg. visualize_cld_graph or any DiagrammeR function.
Use DiagrammeR::generate_dot to create dot output from a DiagrammeR graph object}

\item{file}{Filename - must end in .svg}
}
\value{
Plots dot code and writes .svg file
}
\description{
Save dot code generated in integrater as .svg file
}
