% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dot.R
\name{save_dot}
\alias{save_dot}
\title{Save dot file}
\usage{
save_dot(graphv, filename)
}
\arguments{
\item{graphv}{Graph generated by diagrammR}

\item{filename}{Character vector specifying file name}
}
\value{
Writes a file in dot notation suitable for eg. Atom.
}
\description{
Save dot file
}
\examples{
NA
}
