% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{show_3_motifs}
\alias{show_3_motifs}
\title{Display all two-level motifs consisting of three nodes}
\usage{
show_3_motifs()
}
\value{
Opens image
}
\description{
Display all two-level motifs consisting of three nodes
}
