% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_zoo.R
\name{supported_classes}
\alias{supported_classes}
\title{Lists all supported motif classes for a given signature}
\usage{
supported_classes(signature, directed)
}
\arguments{
\item{signature}{head of a motif identifier string, i.e. string with
comma-separated list specifying the signature of the motif}

\item{directed}{whether the motifs are directed.}
}
\value{
list of supported motif classes
}
\description{
Returns a list with all supported motif classes for the given signature.
Raises an error if the given signature is not supported.
}
\examples{
\dontrun{
supported_classes("1,2", FALSE)
supported_classes("1,1", TRUE)
}
}
\seealso{
\code{supported_signatures()}
}
