% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{toPyGraph}
\alias{toPyGraph}
\title{Translate multi-level statnet network object to networkx (python) object}
\usage{
toPyGraph(g, typeAttr, relabel = TRUE)
}
\arguments{
\item{g}{statnet network object}

\item{typeAttr}{character vector specifying the attribute name where level information is stored in statnet object}

\item{relabel}{should nodes be relabeled with statnet vertex.names? (defaults to TRUE)}
}
\value{
python graph object
}
\description{
Translate multi-level statnet network object to networkx (python) object
}
