% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cld.R
\name{visualize_cld_graph}
\alias{visualize_cld_graph}
\title{Visualize a causal loop diagramm}
\usage{
visualize_cld_graph(
  concepts_df,
  id_column,
  label_column,
  type_column,
  links_df,
  signed = TRUE,
  layout = c("custom_dot", "circo", "dot", "neato", "twopi", "fdp"),
  remove_isolates = FALSE,
  label_size = 14,
  color_mapping = NULL,
  shape_mapping = NULL,
  ranksep = 2
)
}
\arguments{
\item{concepts_df}{A dataframe with min. one column specifying all unique concepts in CLD}

\item{id_column}{Character string specifying based on which column in concepts_df concepts should be linked to links_df}

\item{label_column}{Optional character string specifying column name based on which to label concepts. Defaults to id_column.}

\item{type_column}{Character string specifying the column in concepts_df upon which node type is based. Also used for custom color and shape.}

\item{links_df}{A dataframe with at least a sender and receiver column specifying links between concepts}

\item{signed}{If true, links are colored based on sign (+ or -). This means a column called "sign" needs to be provided in links_df}

\item{layout}{Choose layout algorithm (based on graphviz). Option are "custom_dot", "circo", "dot", "neato", "twopi" or "fdp".
If custom_dot is used, you need to specify at least one of each node types "Activity" and "Target" as source and sink nodes}

\item{remove_isolates}{Should isolates be removed?}

\item{label_size}{Choose label size. Defaults to 14.}

\item{color_mapping}{Optional: Provide a named list of color values based on values in type column. Example: list(type_value = color_value)}

\item{shape_mapping}{Optional: Provide a named list of shape values based on values in type column. Example: list(shape_value = color_value).
All graphviz shape can be chose. Defaults to "circle"}

\item{ranksep}{Rank separation for dot and custom_dot layouts}
}
\value{
For custom_dot layout: A list with dot_code and a plot object. For all other layouts: A DiagrammeR graph object and plot object
}
\description{
Visualize a causal loop diagramm
}
