% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mnet.R
\name{visualize_mnet}
\alias{visualize_mnet}
\title{Visualize a two-level network of actors and issues}
\usage{
visualize_mnet(
  actor_df,
  issue_df,
  actor_links,
  actor_issue_links,
  issue_links,
  viz = c("multinets", "graphviz")
)
}
\arguments{
\item{actor_df}{Data frame containing actor information. Must contain a column named "actor" providing actor id.}

\item{issue_df}{Data frame containing issue information. Must contain a column named "concept" providing issue id.}

\item{actor_links}{Links between actors, must contain columns named "sender" and "receiver"}

\item{actor_issue_links}{Links between actors and issues, must contain columns named "sender" and "receiver"}

\item{issue_links}{Links between issues, must contain columns named "sender" and "receiver"}

\item{viz}{Choose which R package to use for visualization. At present options are "graphviz" (using DiagrammeR, fdp layout)
and "multinets" using the multinets package}
}
\value{
For graphviz: returns list with plot object and dot code. For multinets: returns plot
}
\description{
Visualize a two-level network of actors and issues
}
