% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actor_el}
\alias{actor_el}
\title{Actor edgelist}
\format{A data frame with n rows and n variables:
\describe{
  \item{sender}{tbd}
  \item{receiver}{tbd}
  ...
}}
\source{
survey
}
\usage{
actor_el
}
\description{
A dataset
}
\keyword{datasets}
