% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actors}
\alias{actors}
\title{Actor dataframe}
\format{A data frame with 80 rows and 1 variable:
\describe{
  \item{actor}{Actor names}
  ...
}}
\source{
Reussebene survey
}
\usage{
actors
}
\description{
A dataset specifying the list of unique actors
}
\keyword{datasets}
