% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cld_concepts}
\alias{cld_concepts}
\title{CLD concepts dataframe}
\format{A data frame with 52 rows and 3 variables:
\describe{
  \item{concept}{Short name of the concept}
  \item{type}{Type of concept based on Open Standards methodology}
  \item{english}{Full english name of concept, used for labeling}
  ...
}}
\source{
survey
}
\usage{
cld_concepts
}
\description{
A dataset providing information on all concepts (nodes) in a causal loop diagram elaborated for the Reussebene
}
\keyword{datasets}
