% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{compare_to_baseline}
\alias{compare_to_baseline}
\title{Compares number of motif in a given graph with a random baseline.}
\usage{
compare_to_baseline(
  net,
  motifs,
  n = 10,
  type_attr = c("sesType"),
  assume_sparse = TRUE
)
}
\arguments{
\item{motifs}{list of motif identifier strings}

\item{n}{number of random graphs}

\item{assume_sparse}{whether the random graphs shall be assumed to be sparse.
used to find ideal counting function}

\item{g}{statnet network object}

\item{typeAttr}{character vector specifying the attribute name where level
information is stored in statnet object. The attribute should be a binary
vector. 1 indicates a "social" node and 0 indicates a "non-social" node.}
}
\value{
data frame with one column for each motif identifier string and one
  row for every computed random graph
}
\description{
Compares number of motif in a given graph with a random baseline.
}
\examples{
compare_to_baseline(ml_net, list('1,2[I.C]', '1,2[II.C]'))
}
