% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{compare_to_baseline}
\alias{compare_to_baseline}
\title{Compare empirical network to random baseline}
\usage{
compare_to_baseline(
  net,
  motifs,
  n = 10,
  lvl_attr = c("sesType"),
  assume_sparse = TRUE,
  model = c("erdos_renyi")
)
}
\arguments{
\item{net}{statnet network object}

\item{motifs}{list of motif identifier strings}

\item{n}{number of random graphs}

\item{lvl_attr}{character vector specifying the attribute name where level
information is stored in statnet object.}

\item{assume_sparse}{whether the random graphs shall be assumed to be sparse.
used to find ideal counting function}

\item{model}{baseline model}
}
\value{
data frame with one row for each motif identifier string and one
  row for every computed random graph
}
\description{
This function compares the motif counts in a given network with the motif
counts in a random baseline of motifs.
}
\examples{
compare_to_baseline(ml_net, list('1,2[I.C]', '1,2[II.C]'))
}
