% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{count_motifs}
\alias{count_motifs}
\title{Count multi-level motifs}
\usage{
count_motifs(
  net,
  motifs,
  type_attr = c("sesType"),
  assume_sparse = TRUE,
  omit_total_result = FALSE
)
}
\arguments{
\item{net}{A statnet network object with a node attribute specifying the level of each node}

\item{motifs}{a list of motif identifiers which shall be counted, e.g. list("1,2[I.C]")}

\item{type_attr}{character vector specifying the attribute name where level information is stored in statnet object}

\item{assume_sparse}{whether the network shall be assumed to be sparse (for optimization)}
}
\value{
list indexed by motif identifers giving the motif counts
}
\description{
Count multi-level motifs
}
\examples{
count_motifs(ml_net, type_attr = c("sesType"), motifs=list("1,2[I.C]", "1,2[II.C]", "2,1[I.C]", "2,1[II.C]"))
}
