% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{directed_dummy_net}
\alias{directed_dummy_net}
\title{Two-level directed network dummy example}
\format{
igraph network object
}
\source{
Dummy data
https://gitlab.com/t.seppelt/sesmotifanalyser/-/tree/master/test/data
}
\usage{
directed_dummy_net
}
\description{
Simple igraph network object based on dummy data
}
\keyword{datasets}
