% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_zoo.R
\name{explore_motifs}
\alias{explore_motifs}
\title{Explore the motif zoo interactively in a shiny app}
\usage{
explore_motifs()
}
\value{
Launches a shiny app where all available motifs can be displayed
}
\description{
Explore the motif zoo interactively in a shiny app
}
