% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{motif_summary}
\alias{motif_summary}
\title{Summary for motif counts and distribution (Erdos-Renyi)}
\usage{
motif_summary(net, type_attr = c("sesType"))
}
\arguments{
\item{type_attr}{character vector specifying the attribute name where level
information is stored in statnet object. The attribute should be a binary
vector. 1 indicates a "social" node and 0 indicates a "non-social" node.}

\item{g}{statnet network object}
}
\value{
dataframe with motif counts, expectations and variances for set of
  selected motifs
}
\description{
Summary for motif counts and distribution (Erdos-Renyi)
}
\examples{
motif_summary(ml_net)
}
