% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{motifs_distribution}
\alias{motifs_distribution}
\title{Compute statistical properties of the distribution of motifs in a random
baseline}
\usage{
motifs_distribution(
  net,
  motifs,
  lvl_attr = c("sesType"),
  model = c("erdos_renyi"),
  level = -1,
  omit_total_result = TRUE
)
}
\arguments{
\item{net}{statnet network object}

\item{motifs}{list of motif identifiers describing the motifs whose
distribution shall be analysed}

\item{lvl_attr}{character vector specifying the attribute name where level
information is stored in statnet object.}

\item{model}{model to be used, either erdos_renyi or actors_choice, see sma
documentation}

\item{level}{additional parameter for actors_choice}

\item{omit_total_result}{whether total results shall be omitted}
}
\value{
data frame with one column giving names of motif identifers and two column giving
 expectation and variances per motif
}
\description{
Compute statistical properties of the distribution of motifs in a random
baseline
}
\examples{
motifs_distribution(ml_net, motif = list('1,2[I.C]'))
}
