% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mnet.R
\name{plot_mnet}
\alias{plot_mnet}
\title{Visualize a multi-level network (using ggraph)}
\usage{
plot_mnet(
  net,
  type_attr = c("sesType"),
  layouts = rep("kk", n_levels),
  label = FALSE
)
}
\arguments{
\item{net}{A statnet network object (igraph or tidygraph should also work)}

\item{type_attr}{The name of the categorical node attribute specifying at which level a node is situated}

\item{layouts}{A list of layouts (see ?ggraph::layout_ggraph) for every level eg. for two levels list("auto","circle")}

\item{label}{logical - should nodes be labelled? (defaults to false)}
}
\value{
A ggraph object
}
\description{
Visualize a multi-level network (using ggraph)
}
