% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{show_3_motifs}
\alias{show_3_motifs}
\title{Display all supported motifs with signature ``1,2``.}
\usage{
show_3_motifs()
}
\value{
Opens image
}
\description{
Display all supported motifs with signature ``1,2``.
}
\examples{
show_3_motifs()
}
