% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{show_motif}
\alias{show_motif}
\title{Plots an example for a motif with given motif identifier string taken from
the given graph.}
\usage{
show_motif(net, motif, type_attr = c("sesType"), ...)
}
\arguments{
\item{motif}{motif identifier string for the motif}

\item{...}{additional arguments to be passed to plotting function}

\item{g}{statnet network object}

\item{typeAttr}{character vector specifying the attribute name where level
information is stored in statnet object. The attribute should be a binary
vector. 1 indicates a "social" node and 0 indicates a "non-social" node.}
}
\value{
plot
}
\description{
Plots an example for a motif with given motif identifier string taken from
the given graph.
}
\examples{
show_motif(ml_net, motif = '1,2[I.C]')
}
