% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{show_motif}
\alias{show_motif}
\title{Plots an example for a motif with given motif identifier string taken from
the given graph.}
\usage{
show_motif(motif, net = dummy_net, type_attr = c("sesType"), ...)
}
\arguments{
\item{motif}{motif identifier string for the motif}

\item{net}{statnet network object}

\item{type_attr}{character vector specifying the attribute name where level
information is stored in statnet object. The attribute should be a binary
vector. 1 indicates a "social" node and 0 indicates a "non-social" node.}

\item{...}{additional arguments to be passed to plotting function}
}
\value{
plot
}
\description{
If no network is provided, a motif in a dummy network will be shown.
}
\examples{
show_motif('1,2[I.C]', net = ml_net)
}
\seealso{
exemplify_motif
}
