% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{simulate_baseline}
\alias{simulate_baseline}
\title{Simulate a random baseline}
\usage{
simulate_baseline(
  net,
  motifs,
  n = 10,
  lvl_attr = c("sesType"),
  assume_sparse = TRUE,
  model = "erdos_renyi"
)
}
\arguments{
\item{net}{statnet network object}

\item{motifs}{list of motif identifier strings}

\item{n}{number of random graphs}

\item{lvl_attr}{character vector specifying the attribute name where level
information is stored in statnet object.}

\item{assume_sparse}{whether the random graphs shall be assumed to be sparse.
used to find ideal counting function}

\item{model}{baseline model to be used}
}
\value{
data frame with one column for each motif identifier string and one
  row for every computed random graph
}
\description{
A specified number of random networks using a modified Erdős-Rényi model is
computed. In each of the random networks motifs are counted. A dataframe with
these counts is returned.
}
\examples{
simulate_baseline(ml_net, list('1,2[I.C]'), n = 10)
}
