% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_zoo.R
\name{supported_signatures}
\alias{supported_signatures}
\title{Lists all supported signatures}
\usage{
supported_signatures()
}
\value{
data frame with all supported signatures
}
\description{
Returns a data frame with two columns: signature and a logical value
indicating whether the motifs are directed
}
\examples{
supported_signatures()
}
\seealso{
\code{supported_classes()}
}
