% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidygraph_dummy_net}
\alias{tidygraph_dummy_net}
\title{Two-level tidygraph network example}
\format{
tidygraph network object
}
\source{
Dummy data
 \url{https://github.com/marioangst/motifr/blob/master/notes/tidygraph_dummy_net.R}
}
\usage{
tidygraph_dummy_net
}
\description{
Simple \code{tidygraph} network object for testing
}
\examples{
plot_mnet(tidygraph_dummy_net)
}
\keyword{datasets}
