% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_motifs.R
\name{to_py_graph}
\alias{to_py_graph}
\title{Translate multi-level statnet network object to networkx (python) object}
\usage{
to_py_graph(g, lvl_attr, relabel = TRUE, directed = NULL)
}
\arguments{
\item{g}{statnet network object}

\item{lvl_attr}{character vector specifying the attribute name where level
information is stored in statnet object.}

\item{relabel}{should nodes be relabelled with statnet vertex.names? (defaults
to TRUE)}

\item{directed}{whether the graph shall be treated as a directed graph. Per
default (NULL) this is determined automatically using network::is.directed.}
}
\value{
python graph object
}
\description{
Translate multi-level statnet network object to networkx (python) object
}
\examples{
to_py_graph(motifr::dummy_net, lvl_attr = "sesType")
}
