% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_python.R
\name{update_motifr}
\alias{update_motifr}
\title{Checks for updates for motifr's Python core, the sma package}
\usage{
update_motifr()
}
\description{
It might be necessary to restart your R session after updating the sma package.
}
